Use immutable vector instead of position class and remove position class

+++ SteeringCS/grid/Grid.cs
@@ -173,8 +173,8 @@ namespace SteeringCS.grid
             int halfWallTileSize = wallTile.Size / 2;
             int measureBuffer = 1;
 
-            int wallTileCenterX = wallTile.Position.X + halfWallTileSize;
-            int wallTileCenterY = wallTile.Position.Y + halfWallTileSize;
+            int wallTileCenterX = (int) wallTile.Position.X + halfWallTileSize;
+            int wallTileCenterY = (int) wallTile.Position.Y + halfWallTileSize;
 
             double northDistanceFromWallTileCenter = wallTileCenterY - position.Y + measureBuffer;
             double eastDistanceFromWallTileCenter = position.X - wallTileCenterX + measureBuffer;
@@ -367,8 +367,8 @@ namespace SteeringCS.grid
                 {
                     GridTile gridTile = _gridTiles[x, y];
 
-                    rectangle.X = gridTile.Position.X;
-                    rectangle.Y = gridTile.Position.Y;
+                    rectangle.X = (int) gridTile.Position.X;
+                    rectangle.Y = (int) gridTile.Position.Y;
 
                     if (gridTile is PathTile pathTile && !pathTile.IsEmpty())
                     {
